Add files via upload
(WIP) Dynamic Arrays library in C. This Library is a beginner project to support the dynamic allocation of arrays in C, while abstracting away some of the annoyances. The functions use ptrs to achieve "generic" datatype function, which can be replicated for one of vars by using a function cast for the corresponding datatype. Would like to receive feedback. TODO - When finished compile into a properly downloadable Lib.